feat: 모임 CRUD 헤더 토큰 붙이기
- 방장만 수정, 삭제 가능

+++ src/meeting/meeting.service.ts
@@ -1,17 +1,28 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { Meeting } from './meeting.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { MeetingDTO } from './dto/create-meeting.dto';
+import { AuthService } from 'src/auth/auth.service';
 
 @Injectable()
 export class MeetingService {
   constructor(
     @InjectRepository(Meeting) private meetingRepository: Repository<Meeting>,
+    private readonly authService: AuthService,
   ) {}
 
-  async createMeeting(data: MeetingDTO): Promise<Meeting> {
-    const meeting = this.meetingRepository.create(data);
+  async createMeeting(token: string, data: MeetingDTO): Promise<Meeting> {
+    const { sub } = await this.authService.verifyToken(token);
+
+    const meeting = this.meetingRepository.create({
+      owner_user_id: sub,
+      ...data,
+    });
 
     return await this.meetingRepository.save(meeting);
   }
@@ -35,23 +46,42 @@ export class MeetingService {
     return meeting;
   }
 
-  // 방장인지 확인해야 됨
   async modifyMeeting({
+    token,
     where,
     data,
   }: {
+    token: string;
     where: { id: number };
     data: MeetingDTO;
   }) {
-    await this.findMeeting(where);
+    const { sub } = await this.authService.verifyToken(token);
+
+    const meeting = await this.findMeeting(where);
+    if (meeting.owner_user_id !== sub) {
+      throw new UnauthorizedException('방장만 수정 가능합니다.');
+    }
+
     await this.meetingRepository.update(where, data);
 
     return { success: true };
   }
 
   // 방장인지 확인
-  async removeMeeting(where: { id: number }) {
-    await this.findMeeting(where);
+  async removeMeeting({
+    token,
+    where,
+  }: {
+    token: string;
+    where: { id: number };
+  }) {
+    const { sub } = await this.authService.verifyToken(token);
+
+    const meeting = await this.findMeeting(where);
+    if (meeting.owner_user_id !== sub) {
+      throw new UnauthorizedException('방장만 삭제 가능합니다.');
+    }
+
     await this.meetingRepository.delete(where);
 
     return { success: true };
